fix: Type cast CVPixelBufferGetPlaneCount to Int32 for header
Fixed compilation error where planeCount needed explicit Int32 cast.

+++ NovaidAssist/NovaidAssist/Services/PixelBufferTransmissionService.swift
@@ -23,7 +23,7 @@ class PixelBufferTransmissionService {
             width: Int32(width),
             height: Int32(height),
             pixelFormat: pixelFormat,
-            planeCount: CVPixelBufferGetPlaneCount(pixelBuffer)
+            planeCount: Int32(CVPixelBufferGetPlaneCount(pixelBuffer))
         )
 
         var data = Data()
